requirements.txt
langchain
langchain-community
langgraph
langchain-openai
faiss-cpu